Pattern adjustments

+++ src/patterns/firework.cpp
@@ -14,7 +14,7 @@ void firework(Leds& leds, int32_t timeMs, int periodMs, const CRGBPalette16& pal
     // Original accumulated ~2.5 units/frame via beatsin8(10,1,4); stateless approx: timeMs/8.
     int16_t dist = (int16_t)((uint32_t)timeMs >> 3);
 
-    const auto fireworkFrac      = timeFrac8(timeMs, periodMs);
+    const auto fireworkFrac      = timeFrac8((uint32_t)timeMs, (uint32_t)periodMs);
     const uint8_t fireworkEased  = easeOutQuart((float)fireworkFrac / 255) * 255;
     const uint8_t fireworkLerpVal = lerp8by8(0, leds.size(), fireworkEased);
